backend/requirements.txt
Django==5.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.3.0
pysnmp==6.1.2
ping3==4.0.4
psutil==5.9.8
APScheduler==3.10.4
pytest==8.2.0
pytest-django==4.8.0
google-generativeai==0.8.3
